+++ build-context/package.json
@@ -7,6 +7,7 @@
     "@actions/github": "^5.1.1"
   },
   "devDependencies": {
+    "typescript": "5.1.6",
     "@vercel/ncc": "^0.36.1"
   },
   "scripts": {
