--- build-context/src/index.js
@@ -1,50 +1,0 @@
-const core = require("@actions/core");
-const github = require("@actions/github");
-
-// Context
-const isPullRequest = github.context.eventName === "pull_request";
-const gitSha = isPullRequest
-  ? github.context.payload.pull_request.head.sha
-  : github.sha;
-const gitBranch = isPullRequest
-  ? github.context.payload.pull_request.head.ref
-  : github.ref_name;
-
-// Input
-const shortImageName =
-  core.getInput("short-image-name") || github.context.repo.repo;
-const defaultBranch = core.getInput("default-branch");
-
-// Environment variables
-const projectsEnv = process.ENV["BUILD_CONTEXT_PROJECTS"];
-if (!projectsEnv) {
-  throw new Error("BUILD_CONTEXT_PROJECTS is not set");
-}
-
-const escapedBranch = gitBranch.replaceAll(/[^A-Za-z0-9]/g, "-");
-const testingRegistry = "ghcr.io/wunderflats";
-const projects = projectsEnv.split(",").map((project) => project.trim(project));
-
-// Outputs
-
-core.setOutput("git-sha", gitSha);
-core.setOutput("git-branch", gitBranch);
-core.setOutput("git-escaped-branch", escapedBranch);
-
-for (const project of projects) {
-  const testImageRepository = `${testingRegistry}/${shortImageName}`;
-  const testImageSha = `${testImageRepository}:${gitSha}`;
-  const testImageBranch = `${testImageRepository}:${escapedBranch}`;
-  const testImageDefaultBranch = `${testImageRepository}:${defaultBranch}`;
-  const outputs = [
-    [`${project}-image-repository`, testImageRepository],
-    [`${project}-image-sha`, testImageSha],
-    [`${project}-image-branch`, testImageBranch],
-    [`${project}-image-${defaultBranch}`, defaultBranch],
-  ];
-
-  for (const [key, value] of outputs) {
-    console.log(key, "=", value);
-    core.setOutput(key, value);
-  }
-}
