+++ build-context/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+      "module": "es2022",
+      "target": "es2022",
+      "moduleResolution": "nodenext"
+    }
+}
